stock_quantity check is now nested within the validation for the user's quantity request. If user request is > stock_quantity, it logs a message and re-asks the question.

+++ bamazon.sql
@@ -7,8 +7,8 @@ CREATE TABLE products (
 	item_id INT AUTO_INCREMENT NOT NULL,
     product_name TEXT (100) NOT NULL,
 	department_name TEXT (100) NOT NULL,
-    price DECIMAL (10, 4) NOT NULL,
-    stock_quantity DECIMAL (10, 4) NOT NULL,
+    price DECIMAL (10, 2) NOT NULL,
+    stock_quantity INT NOT NULL,
     PRIMARY KEY (item_id)
 );
